Transforming the header to source and header!
