+++ install.py
@@ -13,19 +13,38 @@ from shutil import rmtree
 QEasyDownloader = {
         "username" : "antony-jr",
         "repo"     : "QEasyDownloader",
-        "mkdir"    : "QEasyDownloader",
+        "mkdir"    : {
+            "QEasyDownloader/src",
+            "QEasyDownloader/include"
+        },
         "install"  : {
-            "QEasyDownloader.hpp" : "QEasyDownloader/QEasyDownloader.hpp",
+            "CMakeLists.txt"       : "QEasyDownloader/CMakeLists.txt",
+            "QEasyDownloader.pro"         : "QEasyDownloader/QEasyDownloader.pro",
+            "QEasyDownloader"             : "QEasyDownloader/QEasyDownloader",
+            "include/QEasyDownloader.hpp" : "QEasyDownloader/include/QEasyDownloader.hpp",
+            "src/QEasyDownloader.cc"      : "QEasyDownloader/src/QEasyDownloader.cc",
             "LICENSE"      : "QEasyDownloader/LICENSE"
         }
 }
 
 def installPackage(config):
     print("Installing " + config["repo"])
-    print("Creating Directory " + config["mkdir"])
-    if os.path.exists(config["mkdir"]):
-        rmtree(config["mkdir"])
-    os.mkdir(config["mkdir"]) # Make the directory!
+    # Make parent directory first.
+    if os.path.isfile(config["repo"]):
+        print("Deleting duplicate file(s)... ")
+        os.remove(config["repo"])
+        os.mkdir(config["repo"])
+    else:
+     if os.path.exists(config["repo"]):
+         rmtree(config["repo"])
+     else:
+         os.mkdir(config["repo"])
+
+    for i in config["mkdir"]:
+     print("Creating Directory " + i)
+     if os.path.exists(i):
+         rmtree(i)
+     os.mkdir(i) # Make the directory!
     print("Downloading the latest release from github... ")
 
     # Write files from the repo
@@ -36,7 +55,7 @@ def installPackage(config):
             fp.write(it)
         fp.close()
 
-    print("Installed "+config["repo"] + ".")
+    print("Installed "+config["repo"]+".")
     return True
 
 if __name__ == "__main__":
